Added a paragraph on related work
It's heavily oriented towards human movement, but I added a statement that this is just one area of application. Maybe this text is already too much.  If we try to do a complete review, it becomes easy to criticize the effort for not being complete...  My attempt at feature comparison will be similarly flawed.

Before I add the references in the bib file, let me know if this is the right direction.