Enables support for additional path constraints in the EoMs.
Fixes #438

Problem and ConstraintCollocator can now except M equations of motion
that != n states. M = n was a hard requirement but it is unnecessary and
prevents adding additional constraint equations that are valid for all
time nodes. This is fixed by added a new attribute num_eom and adjusting
the various functions that output the constraints to use it rather than
num_states. A beginner example showing this works for a particle moving
on a circle is included to demonstrate that additional path constraints
works.