Support for inequality constraints.
Fixes #156

New kwargs added to Problem that allow a user to set lower and upper
bounds for each equation in the equations of motion. This allows
inequality path constraints to be added to the essential differential
equations. Caution should be used when added lower and upper bounds
other than 0 to the essential differential equations, but it is
supported.