Enables caching of the compiled code if tmp_dir is set.
Fixes #107

A hash of the unique evaluation code is stored in its C file and the
files present in the tmp_dir are checked to see if they contain the
hash. If one does, then that compiled module is loaded if possible. This
saves having to recompile the C code if the differential algebraic
equations passed to Problem or ConstraintCollocator have not changed.